rename hacker_news_assets package name to avoid conflicts
