+++ examples/hacker_news_assets/hacker_news_assets/resources/dbt_asset_resource.py
