+++ examples/hacker_news_assets/hacker_news_assets/resources/fixed_s3_pickle_io_manager.py
