+++ examples/hacker_news_assets/hacker_news_assets/resources/hn_resource.py
