+++ examples/hacker_news_assets/hacker_news_assets/sensors/download_pipeline_finished_sensor.py
