+++ examples/hacker_news_assets/hacker_news_assets/solids/recommender_model.py
@@ -3,7 +3,7 @@ import random
 
 from dagster import EventMetadata, Output
 from dagster.core.asset_defs import AssetIn, asset
-from hacker_news.solids.user_story_matrix import IndexedCooMatrix
+from hacker_news_assets.solids.user_story_matrix import IndexedCooMatrix
 from pandas import DataFrame, Series
 from sklearn.decomposition import TruncatedSVD
