+++ examples/hacker_news_assets/hacker_news_assets/utils/snapshot.py
@@ -3,7 +3,7 @@ import json
 from concurrent.futures import ThreadPoolExecutor
 
 from dagster.utils import file_relative_path
-from hacker_news.resources.hn_resource import HNAPIClient
+from hacker_news_assets.resources.hn_resource import HNAPIClient
 from tqdm import tqdm
 
 # Slice that surrounds 12/30/2020
