+++ examples/hacker_news_assets/hacker_news_assets_tests/test_pipelines/test_story_recommender.py
@@ -1,0 +1,2 @@
+# pylint: disable=unused-import
+from hacker_news_assets.pipelines.story_recommender import story_recommender_dev, story_recommender_prod
