+++ examples/hacker_news_assets/hacker_news_assets_tests/test_solids/test_comment_stories.py
@@ -1,5 +1,5 @@
 import pytest
-from hacker_news.solids.comment_stories import comment_stories
+from hacker_news_assets.solids.comment_stories import comment_stories
 from pandas import DataFrame
