+++ examples/hacker_news_assets/hacker_news_assets_tests/test_solids/test_user_story_matrix.py
@@ -1,5 +1,5 @@
 import pytest
-from hacker_news.solids.user_story_matrix import user_story_matrix
+from hacker_news_assets.solids.user_story_matrix import user_story_matrix
 from pandas import DataFrame
