+++ examples/hacker_news_assets/hacker_news_assets_tests/test_solids/test_user_top_recommended_stories.py
@@ -1,6 +1,6 @@
 import numpy as np
-from hacker_news.solids.user_story_matrix import IndexedCooMatrix
-from hacker_news.solids.user_top_recommended_stories import user_top_recommended_stories
+from hacker_news_assets.solids.user_story_matrix import IndexedCooMatrix
+from hacker_news_assets.solids.user_top_recommended_stories import user_top_recommended_stories
 from pandas import DataFrame, Series
 from scipy.sparse import coo_matrix
 from sklearn.decomposition import TruncatedSVD
