--- examples/hacker_news_assets/hacker_news_tests/test_pipelines/test_story_recommender.py
@@ -1,2 +1,0 @@
-# pylint: disable=unused-import
-from hacker_news.pipelines.story_recommender import story_recommender_dev, story_recommender_prod
