+++ examples/hacker_news_assets/tox.ini
@@ -28,11 +28,11 @@ whitelist_externals =
   pylint
 commands =
   echo -e "--- \033[0;32m:lint-roller: Running pylint\033[0m"
-  pylint -j 0 --rcfile=../../.pylintrc hacker_news hacker_news_tests
+  pylint -j 0 --rcfile=../../.pylintrc hacker_news_assets hacker_news_assets_tests
 
 [testenv:mypy]
 whitelist_externals =
   mypy
 commands =
   echo -e "--- \033[0;32m:mypy: Running mypy\033[0m"
-  mypy -p hacker_news -p hacker_news_tests --namespace-packages --ignore-missing-imports
+  mypy -p hacker_news_assets -p hacker_news_assets_tests --namespace-packages --ignore-missing-imports
