+++ examples/hacker_news_assets/workspace.yaml
@@ -1,2 +1,2 @@
 load_from:
-  - python_package: hacker_news
+  - python_package: hacker_news_assets
